+++ allure.go
@@ -44,7 +44,7 @@ const (
 	stepDeadlineWindow  = 75 * time.Millisecond
 )
 
-//go:generate go tool ifacemaker -f $GOFILE -o interface.go -s PluginAllure -i Interface -p $GOPACKAGE -e Plugin -y "Interface defines allure plugin interface.\nUseful for writing helpers which require allure methods but can't rely on concrete type." -x -e panicked -e status -e asResult -e parameters -e links -e attachments -e allRawAttachments -e title -e asStep -e timeBoundaries -e steps -e containers -e beforeEach -e afterEach -e hooks -e addMessage -e addTrace -e overrides -e results -e resultsGroupParametrized -e afterAll -e writeResults -e writeContainers -e writeAttachments -e writeAttachment -e writeProperties -e writeCategories -e labels -e attachmentPath -e baseName -e testCaseID -e historyID -e resultsFlattenParametrized -e statusDetails -e suiteName -e plugin -e beforeAll -e cleanup -e writeReport -e plan -e applyOptions -e fullName -e createOutputDir -e asContainer -e beforeEachSub -e afterEachSub -e propagatedStatusDetails -e hookDescendants -e descendants -e testChildren -e hasTestNeighbors -e subtest -e attach -e parentSuiteName -e realStatus
+//go:generate go tool ifacemaker -f $GOFILE -o interface.go -s PluginAllure -i Interface -p $GOPACKAGE -e Plugin -y "Interface defines allure plugin interface.\nUseful for writing helpers which require allure methods but can't rely on concrete type." -x -e panicked -e status -e asResult -e parameters -e links -e attachments -e allRawAttachments -e title -e asStep -e timeBoundaries -e steps -e containers -e beforeEach -e afterEach -e hooks -e addMessage -e addTrace -e overrides -e results -e resultsGroupParametrized -e afterAll -e writeResults -e writeContainers -e writeAttachments -e writeAttachment -e writeProperties -e writeCategories -e labels -e attachmentPath -e baseName -e testCaseID -e historyID -e resultsFlattenParametrized -e statusDetails -e suiteName -e plugin -e beforeAll -e cleanup -e writeReport -e plan -e applyOptions -e fullName -e createOutputDir -e asContainer -e beforeEachSub -e afterEachSub -e propagatedStatusDetails -e hookDescendants -e descendants -e testChildren -e hasTestNeighbors -e subtest -e attach -e parentSuiteName -e realStatus -e packageName
 
 var _ Interface = (*PluginAllure)(nil)
 
@@ -1682,6 +1682,7 @@ func (a *PluginAllure) labels() []Label {
 	for _, l := range []Label{
 		{Name: labelParentSuite, Value: a.parentSuiteName()},
 		{Name: labelSuite, Value: a.suiteName()},
+		{Name: labelPackage, Value: a.packageName()},
 		{Name: labelHost, Value: hostname},
 		{Name: labelLanguage, Value: "go"},
 		{Name: labelFramework, Value: "testo"},
@@ -1781,6 +1782,29 @@ func (a *PluginAllure) suiteName() string {
 	}
 
 	return s.Name
+}
+
+func (a *PluginAllure) packageName() string {
+	var pc uintptr
+
+	switch test := testo.Reflect(a).Test.(type) {
+	case testoreflect.ParametrizedTestInfo:
+		pc = test.FuncPC
+
+	case testoreflect.RegularTestInfo:
+		pc = test.FuncPC
+
+	default:
+		return ""
+	}
+
+	funcName := runtime.FuncForPC(pc).Name()
+
+	lastSlash := max(0, strings.LastIndexByte(funcName, '/'))
+
+	dotAfterLastSlash := strings.IndexByte(funcName[lastSlash:], '.') + lastSlash
+
+	return funcName[:dotAfterLastSlash]
 }
 
 func newProperties() properties {